[FIX] website, *: fix `scrollTo` and `closestScrollableY` issues
*: web_editor, website_blog, website_forum

In commit [1], the `dom.js` functions linked to the scroll were
converted in order to not use jQuery anymore. They were then later
completely removed from this file in commit [2] and then added in
`snippets.editor.js`, as well as some other functions that would be
needed in this file, as it was considered that only this file needed the
old implementation. Every other call to `scrollTo` was then replaced by
a call to a new implementation made in the `scrolling.js` utils. The
same goes for the calls to `closestScrollable` which were replaced by
the new `closestScrollableY`.

However, this should not have been done, as any scroll occurring in a
website page should always consider the header (dis)appearing during the
scroll, the `#wrapwrap` as scrollable element, as well as other website
specific behaviors. A lot of behaviors were therefore broken all across
website pages, leading to tracebacks for some of them.

This commit therefore fixes these issues by using the old implementation
again where it is necessary. The web_editor `scrolling.js` utils file is
therefore created and imported instead of the web `scrolling.js` when
needed. It contains the converted functions of commit [2]. Note that the
`scrollTo` function was fixed in the previous commit of this PR, to
consider the header, among other things. This commit then further fixes
it, so it can consider correctly the anchors strings (`#top` and
`#bottom`).

The issues fixed by this PR are the following:
For all issues, first drop enough snippets to have a scrollbar.

1) Issue when moving snippets with the arrows:
- Click on a snippet and move it by using the overlay arrow.
=> The page scroll is not smooth and if we clicked on the top arrow, it
does not take into account the header, which can therefore cover it.

2) Issue with the "Background Position" option:
- Drop the "Big Boxes" or "Masonry" snippets at a place where we can
hide it by scrolling downwards.
- Add a background image on a column.
- Scroll down to hide the snippet and then click on the "Background
Position" button to adjust the image position.
=> The page scroll does not take into account the header and the extra
offset.

3) Issue with the image cropper:
- Click on an image, then hide it by scrolling.
- Click on the "Crop" option.
=> The page scroll should bring the cropper at the top of the page, just
below the header, which is not the case.
Note that for this one, the `closestScrollableY` was okay, since this
function implementation was in fact taken from here, so it is the same
as before (see commit [3]).

4) Issue with anchors:
- Set a button link to `#top` or `#bottom` then save.
- Click on the buttons.
=> It does not work anymore, because what is given to the `scrollTo`
function is a string in that case, and not an element.

5) Issue with the "Height" option with dynamic snippets:
- On an empty page, drop a dynamic snippet (e.g. "Products").
- Set the "Height" option to 100%.
=> Traceback because of `closestScrollableY`.

6) Issue with the "Table of Content" snippet:
- Drop a "Table of Content" snippet.
- Hide it on desktop with the "Visibility" option then save.
=> Traceback because of `closestScrollableY`.
- Drop "Table of Content" inside a "Popup" snippet and save.
=> Traceback.

7) Issue when submitting a form:
- Drop the "Form" snippet and redirect the success page to `#top`.
- Save, then fill the form and submit it.
=> It does not scroll.

8) Issue with comments on blog posts:
- On a blog post, enable the comments.
- Activate the "Select to Comment" option and then save.
- Select some text in the post and click on the comment popover.
=> The scroll is not smooth.

9) Issue with blog posts "scroll arrows":
- On a blog post, click on the arrow of the banner or on the "next post"
arrow at the bottom.
=> Traceback. It happens because the `duration` argument was removed
from the `_forumScrollAction` function but not from the calls to this
function in commit [1]. Once that fixed, the scroll still does not work
correctly, as it does not consider the header offset when scrolling. For
example, it ends after the breadcrumbs, while they should be visible.

10) Issue when replying to a Forum post:
- Go to a Forum post with no answers yet.
- Scroll to the bottom of the page and then click on the "Reply" button.
=> The page should scroll to the answer "popup", which is not the case.

[1]: https://github.com/odoo/odoo/commit/8c5a7dc788488baccf42f1a99d1e9107c03083ef
[2]: https://github.com/odoo/odoo/commit/77797a51bbbdc8d62ba44a984d546823f23f935a
[3]: https://github.com/odoo/odoo/commit/429f9792d87892a4f5ea402ef8750bcedbe3a0d8

task-4128905
