+++ addons/website_blog/static/src/js/contentshare.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
 import { sprintf, escape } from "@web/core/utils/strings";
-import { scrollTo } from "@web/core/utils/scrolling";
+import { scrollTo } from "@web_editor/js/common/scrolling";
 
 export function share(el, options) {
     const option = {
